requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-magic-bin==0.4.14
aiofiles==23.2.1
pdfplumber==0.10.3
python-docx==1.1.0
pillow==10.1.0
